+++ src/components/SliderComponent/SliderComponent.scss
@@ -24,9 +24,8 @@
 
 .slick-track {
   display: flex;
-  gap: 3rem;
+  gap: 1rem;
   padding: 1rem 0;
-  background-color: rgba(0, 0, 0, 0.027);
 }
 
 .slider__text {
