+++ src/pages/BuildMacroPage/BuildMacroPage.jsx
@@ -1,5 +1,9 @@
 import "./BuildMacroPage.scss";
 
-export default function BuildMacroPage() {
-  return <div className="macro-page"></div>;
+export default function BuildMacroPage({ userProfile, loginState }) {
+  return (
+    <div className="macro-page">
+      <h2>Follow these below</h2>
+    </div>
+  );
 }
